Agentic AI Projects/Project 3/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pymongo==4.6.0
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
google-generativeai==0.3.2
langgraph==0.0.20
langchain==0.1.0
langchain-google-genai==0.0.6
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pillow==10.1.0
requests==2.31.0
schedule==1.2.0
pandas==2.1.4
numpy==1.26.2
python-dateutil==2.8.2

